+++ cuda/src/kernels/array/pad.rs
@@ -89,12 +89,12 @@ impl Pad {
         let mut launch_args = TractLaunchArgs::new(stream, &func);
         launch_args.set_view(&i_view);
         launch_args.set_view(&o_view);
-        launch_args.set_slice::<i64>(&in_shape);
-        launch_args.set_slice::<i64>(&out_shape);
-        launch_args.set_slice::<i64>(&in_strides);
-        launch_args.set_slice::<i64>(&pad_before);
+        launch_args.set_slice::<i32>(&in_shape);
+        launch_args.set_slice::<i32>(&out_shape);
+        launch_args.set_slice::<i32>(&in_strides);
+        launch_args.set_slice::<i32>(&pad_before);
         launch_args.set_el::<T>(*fill_value);
-        launch_args.set_el::<i64>(len);
+        launch_args.set_el::<i32>(len);
 
         let cfg = LaunchConfig::for_num_elems(len as _);
         launch_args.launch(cfg)
